python/packages/kagent-rag-processor/pyproject.toml
[project]
name = "kagent-rag-processor"
version = "0.1.0"
description = "RAG document processor for kagent - extracts text, creates embeddings, stores in pgvector"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{ name = "kagent-dev", email = "kagent@kagent.dev" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "minio>=7.2.0",
    "psycopg2-binary>=2.9.0",
    "sentence-transformers>=3.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "pandas>=2.2.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
kagent-rag-processor = "kagent.rag_processor.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kagent"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
